Added Searching Algorithms in Java
Linear Search algorithm searches through an array and returns the index of the value it searches for if present

Binary search algorithm compares the target value to the middle element of the array. If they are not equal, the half in which the target cannot lie is eliminated and the search continues on the remaining half.